requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiopg==1.4.0
aiosignal==1.3.2
alembic==1.15.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
frozenlist==1.5.0
graphql-core==3.2.6
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
pika==1.3.2
propcache==0.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
